% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMeta.R
\name{addMeta}
\alias{addMeta}
\title{Add a metadata to be included in the shiny app}
\usage{
addMeta(scConf, meta.to.add, obj)
}
\arguments{
\item{scConf}{shinycell config data.table}

\item{meta.to.add}{metadata to add from the single-cell metadata.
Must match the column names in 
\code{seu@meta.data} i.e. \code{colnames(seu@meta.data)} for 
Seurat objects or the column names in 
\code{sce@colData} i.e. \code{colnames(sce@colData)} for 
SingleCellExperiment objects.}

\item{obj}{original single-cell data object used to create shinycell 
config data.table}
}
\value{
updated shinycell config data.table
}
\description{
Add a metadata to be included in the shiny app.
}
\examples{
scConf = addMeta(scConf, c("orig.ident"), seu)

}
\author{
John F. Ouyang
}
