% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createConfig.R
\name{createConfig}
\alias{createConfig}
\title{Create a shinycell config data.table}
\usage{
createConfig(obj, meta.to.include = NA, legendCols = 4, maxLevels = 50)
}
\arguments{
\item{obj}{input single-cell data object. Both Seurat objects (v3+) and 
SingleCellExperiment objects are accepted.}

\item{meta.to.include}{columns to include from the single-cell metadata. 
Default is \code{NA}, which is to use all columns. Users can specify 
columns to include, which must match the column names in 
\code{seu@meta.data} i.e. \code{colnames(seu@meta.data)} for 
Seurat objects or the column names in 
\code{sce@colData} i.e. \code{colnames(sce@colData)} for 
SingleCellExperiment objects.}

\item{legendCols}{maximum number of columns allowed when displaying the 
legends of categorical metadata}

\item{maxLevels}{maximum number of levels allowed for categorical metadata.
Metadata with nlevels > maxLevels will be discarded automatically}
}
\value{
shinycell config data.table
}
\description{
Create a shinycell config data.table containing (i) the single-cell 
metadata to display on the Shiny app, (ii) ordering of factors / 
categories of categorical metadata and (iii) colour palettes associated 
with each metadata.
}
\examples{
scConf = createConfig(obj)

}
\author{
John F. Ouyang
}
