% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delMeta.R
\name{delMeta}
\alias{delMeta}
\title{Remove a metadata from being included in the shiny app}
\usage{
delMeta(scConf, meta.to.del)
}
\arguments{
\item{scConf}{shinycell config data.table}

\item{meta.to.del}{metadata to delete. Users can either use the original 
metadata column names or display names. For more information regarding 
display name, see \code{?modMetaName}. Multiple metadata can be specified.}
}
\value{
updated shinycell config data.table
}
\description{
Remove a metadata from being included in the shiny app.
}
\examples{
scConf = delMeta(scConf, c("orig.ident"))

}
\author{
John F. Ouyang
}
