% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExampleData.R
\name{getExampleData}
\alias{getExampleData}
\title{Download example Seurat objects}
\usage{
getExampleData(type = c("single", "multi"))
}
\arguments{
\item{type}{can be either "single" or "multi" which downloads one or two 
Seurat objects respectively for the Quick Start/Detailed Tutorial or 
Multi-dataset Tutorial respectively}
}
\value{
downloaded Seurat object
}
\description{
Download example Seurat objects required for ShinyCell tutorial.
}
\examples{
getExampleData()

}
\author{
John F. Ouyang
}
