% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeShinyApp.R
\name{makeShinyApp}
\alias{makeShinyApp}
\title{Make a shiny app}
\usage{
makeShinyApp(
  obj,
  scConf,
  gex.assay = NA,
  gex.slot = c("data", "scale.data", "counts"),
  gene.mapping = NA,
  shiny.title = "scRNA-seq shiny app",
  shiny.footnotes = "\\"\\"",
  shiny.dir = "shinyApp/",
  default.gene1 = NA,
  default.gene2 = NA,
  default.multigene = NA,
  default.dimred = c("UMAP_1", "UMAP_2")
)
}
\arguments{
\item{obj}{input single-cell data object. Both Seurat objects (v3+) and 
SingleCellExperiment objects are accepted.}

\item{scConf}{shinycell config data.table}

\item{gex.assay}{assay in single-cell data object to use for plotting 
gene expression. Default is to either use the "RNA" assay for Seurat 
objects (v3+) or the "logcounts" assay for SingleCellExperiment objects}

\item{gex.slot}{slot in single-cell assay to plot. This is only used 
for Seurat objects (v3+). Default is to use the "data" slot}

\item{gene.mapping}{specify a mapping to convert gene identifiers. A 
named vector must be supplied where \code{names(gene.mapping)} correspond 
to the actual gene identifiers in the gene expression matrix whereas 
\code{gene.mapping} correspond to new identifiers to map to. Partial 
mapping is allowed where the mapping is only provided for some gene 
identifiers, the remaining gene identifiers will remain unchanged and a 
warning message will be presented. Default is \code{NA} which is not to 
perform any mapping}

\item{shiny.title}{title for shiny app}

\item{shiny.footnotes}{text for shiny app footnote. Can be used to insert 
citations for the single-cell data}

\item{shiny.dir}{specify directory to create the shiny app in. Default is 
to create a new directory named "shinyApp"}

\item{default.gene1}{specify primary default gene to show}

\item{default.gene2}{specify secondary default gene to show}

\item{default.multigene}{character vector specifying the default genes to 
show in bubbleplot / heatmap}

\item{default.dimred}{character vector specifying the two default dimension 
reductions}
}
\value{
directory containing shiny app
}
\description{
Make a shiny app based on the shinycell config data.table and single-cell 
data object.
}
\examples{
footnote = paste0(
  'strong("Reference: "), "Liu X., Ouyang J.F., Rossello F.J. et al. ",',
  'em("Nature "), strong("586,"), "101-107 (2020) ",',
  'a("doi:10.1038/s41586-020-2734-6",',
  'href = "https://www.nature.com/articles/s41586-020-2734-6",',
  'target="_blank"), style = "font-size: 125\%;"'
)
makeShinyApp(seu, scConf, 
             shiny.title = "scRNA-seq Shiny app",
             shiny.dir = "shinyApp/", shiny.footnotes = footnote,
             default.gene1 = "NANOG", default.gene2 = "DNMT3L",
             default.multigene = c("ANPEP","NANOG","ZIC2","NLGN4X","DNMT3L",
                                   "DPPA5","SLC7A2","GATA3","KRT19")) 

}
\author{
John F. Ouyang
}
