% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeShinyCodes.R
\name{makeShinyCodes}
\alias{makeShinyCodes}
\title{Generate code files required for shiny app (one dataset)}
\usage{
makeShinyCodes(shiny.title, shiny.footnotes, shiny.prefix, shiny.dir)
}
\arguments{
\item{shiny.title}{title for shiny app}

\item{shiny.footnotes}{text for shiny app footnote}

\item{shiny.prefix}{specify file prefix}

\item{shiny.dir}{specify directory to create the shiny app in}
}
\value{
server.R and ui.R required for shiny app
}
\description{
Generate code files required for shiny app containing only one dataset. In 
particular, two R scripts will be generated, namely \code{server.R} and 
\code{ui.R}. If users want to include multiple dataset in one shiny app, 
please use \code{makeShinyCodesMulti()} instead. Note that both 
\code{makeShinyFiles} and \code{makeShinyCodes} functions are ran when 
running the wrapper function \code{makeShinyApp}.
}
\examples{
makeShinyCodes(shiny.title = "scRNA-seq shiny app", shiny.footnotes = '""',
               shiny.prefix = "sc1", shiny.dir = "shinyApp/")

}
\author{
John F. Ouyang
}
