% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeShinyCodesMulti.R
\name{makeShinyCodesMulti}
\alias{makeShinyCodesMulti}
\title{Generate code files required for shiny app (multi datasets)}
\usage{
makeShinyCodesMulti(
  shiny.title,
  shiny.footnotes,
  shiny.prefix,
  shiny.headers,
  shiny.dir
)
}
\arguments{
\item{shiny.title}{specify the overall title for shiny app}

\item{shiny.footnotes}{text for shiny app footnote}

\item{shiny.prefix}{specify file prefix for each dataset. Must match the 
prefix used in \code{makeShinyFiles}}

\item{shiny.headers}{specify the tab header names for each dataset. Length 
must match that of \code{shiny.prefix}}

\item{shiny.dir}{specify directory to create the shiny app in}
}
\value{
server.R and ui.R required for shiny app
}
\description{
Generate code files required for shiny app containing multiple datasets. In 
particular, two R scripts will be generated, namely \code{server.R} and 
\code{ui.R}. Note that \code{makeShinyFiles} has to be ran prior to 
generate the necessary data files for each of the dataset to be included. 
The prefix used in \code{makeShinyFiles} have to be then supplied in this 
function.
}
\examples{
makeShinyCodes(shiny.title = "scRNA-seq shiny app", shiny.footnotes = '""',
               shiny.prefix = c("sc1", "sc2"), 
               shiny.headers = c("dataset1", "dataset2"),
               shiny.dir = "shinyApp/")
               
}
\author{
John F. Ouyang
}
