% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeShinyFiles.R
\name{makeShinyFiles}
\alias{makeShinyFiles}
\title{Generate data files required for shiny app}
\usage{
makeShinyFiles(
  obj,
  scConf,
  gex.assay = NA,
  gex.slot = c("data", "scale.data", "counts"),
  gene.mapping = NA,
  shiny.prefix,
  shiny.dir = "shinyApp/",
  default.gene1 = NA,
  default.gene2 = NA,
  default.multigene = NA,
  default.dimred = c("UMAP_1", "UMAP_2")
)
}
\arguments{
\item{obj}{input single-cell data object. Both Seurat objects (v3+) and 
SingleCellExperiment objects are accepted.}

\item{scConf}{shinycell config data.table}

\item{gex.assay}{assay in single-cell data object to use for plotting 
gene expression. Default is to either use the "RNA" assay for Seurat 
objects (v3+) or the "logcounts" assay for SingleCellExperiment objects}

\item{gex.slot}{slot in single-cell assay to plot. This is only used 
for Seurat objects (v3+). Default is to use the "data" slot}

\item{gene.mapping}{specify a mapping to convert gene identifiers. A 
named vector must be supplied where \code{names(gene.mapping)} correspond 
to the actual gene identifiers in the gene expression matrix whereas 
\code{gene.mapping} correspond to new identifiers to map to. Partial 
mapping is allowed where the mapping is only provided for some gene 
identifiers, the remaining gene identifiers will remain unchanged and a 
warning message will be presented. Default is \code{NA} which is not to 
perform any mapping}

\item{shiny.prefix}{specify file prefix}

\item{shiny.dir}{specify directory to create the shiny app in}

\item{default.gene1}{specify primary default gene to show}

\item{default.gene2}{specify secondary default gene to show}

\item{default.multigene}{character vector specifying default genes to 
show in bubbleplot / heatmap}

\item{default.dimred}{character vector specifying two default dim. reduction}
}
\value{
data files required for shiny app
}
\description{
Generate data files required for shiny app. Five files will be generated, 
namely (i) the shinycell config \code{prefix_conf.rds}, (ii) the gene 
mapping object config \code{prefix_gene.rds}, (iii) the single-cell gene 
expression \code{prefix_gexpr.h5}, (iv) the single-cell metadata 
\code{prefix_meta.rds} and (v) the defaults for the Shiny app 
\code{prefix_def.rds}. A prefix is specified in each file to allow for the 
use of multiple single-cell datasets in one Shiny app. Note that both 
\code{makeShinyFiles} and \code{makeShinyCodes} functions are ran when 
running the wrapper function \code{makeShinyApp}.
}
\examples{
makeShinyFiles(seu, scConf, gex.assay = "RNA", gex.slot = "data",
               shiny.prefix = "sc1", shiny.dir = "shinyApp/",
               default.gene1 = "GATA3", default.gene2 = "DNMT3L",
               default.multigene = c("ANPEP","NANOG","ZIC2","NLGN4X","DNMT3L",
                                     "DPPA5","SLC7A2","GATA3","KRT19"),
               default.dimred = c("UMAP_1", "UMAP_2")

}
\author{
John F. Ouyang
}
