% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrSVload}
\alias{wrSVload}
\title{Write code for loading objects for server.R}
\usage{
wrSVload(prefix)
}
\arguments{
\item{prefix}{file prefix}
}
\description{
Write code for loading objects for server.R
}
