% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrUIga}
\alias{wrUIga}
\title{Write code for google-analytics.html}
\usage{
wrUIga(gaID)
}
\arguments{
\item{gaID}{Google analytics tracking ID (e.g. "UA-123456789-0")}
}
\description{
Write code for google-analytics.html
}
