% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrUImain}
\alias{wrUImain}
\title{Write code for main block of ui.R}
\usage{
wrUImain(prefix, subst = "", ptsiz = "1.25")
}
\arguments{
\item{prefix}{file prefix}
}
\description{
Write code for main block of ui.R
}
