% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrUIsingle}
\alias{wrUIsingle}
\title{Write code for front portion of ui.R}
\usage{
wrUIsingle(title, ganalytics)
}
\arguments{
\item{title}{shiny app title}

\item{ganalytics}{Google analytics tracking ID (e.g. "UA-123456789-0")}
}
\description{
Write code for front portion of ui.R
}
