% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExampleData.R
\name{getExampleData}
\alias{getExampleData}
\title{Download example Seurat objects / single-cell data}
\usage{
getExampleData(type = c("single", "multi", "h5ad", "loom", "plaintext"))
}
\arguments{
\item{type}{can be either "single" or "multi" or "h5ad" or "loom" 
or "plaintext"}
}
\value{
downloaded Seurat object
}
\description{
Download example Seurat objects / single-cell data required for 
ShinyCell tutorials.
}
\examples{
getExampleData()

}
\author{
John F. Ouyang
}
