% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeShinyCodes.R
\name{makeShinyCodes}
\alias{makeShinyCodes}
\title{Generate code files required for shiny app (one dataset)}
\usage{
makeShinyCodes(
  shiny.title,
  shiny.footnotes,
  shiny.prefix,
  shiny.dir,
  enableSubset = TRUE,
  defPtSiz = 1.25,
  ganalytics = NA
)
}
\arguments{
\item{shiny.title}{title for shiny app}

\item{shiny.footnotes}{text for shiny app footnote. When given as a list, 
citation can be inserted by specifying author, title, journal, volume, 
page, year, doi, link. See example below.}

\item{shiny.prefix}{specify file prefix}

\item{shiny.dir}{specify directory to create the shiny app in}

\item{enableSubset}{specify whether to enable "Toggle to subset cells" 
functionality in the shiny app. Default is to enable this functionality}

\item{defPtSiz}{specify default point size for single cells. For example, a 
smaller size can be used if you have many cells in your dataset}

\item{ganalytics}{Google analytics tracking ID (e.g. "UA-123456789-0")}
}
\value{
server.R and ui.R required for shiny app
}
\description{
Generate code files required for shiny app containing only one dataset. In 
particular, two R scripts will be generated, namely \code{server.R} and 
\code{ui.R}. If users want to include multiple dataset in one shiny app, 
please use \code{makeShinyCodesMulti()} instead. Note that both 
\code{makeShinyFiles} and \code{makeShinyCodes} functions are ran when 
running the wrapper function \code{makeShinyApp}.
}
\examples{
# Example citation
citation = list(
  author  = "Liu X., Ouyang J.F., Rossello F.J. et al.",
  title   = "",
  journal = "Nature",
  volume  = "586",
  page    = "101-107",
  year    = "2020", 
  doi     = "10.1038/s41586-020-2734-6",
  link    = "https://www.nature.com/articles/s41586-020-2734-6")
makeShinyCodes(shiny.title = "scRNA-seq shiny app", shiny.footnotes = "",
               shiny.prefix = "sc1", shiny.dir = "shinyApp/")

}
\author{
John F. Ouyang
}
